+++ Source/Tools/LinuxEmulation/LinuxSyscalls/ThreadManager.h
@@ -8,7 +8,7 @@ $end_info$
 
 #pragma once
 
-#include "Common/Profiler.h"
+#include "Common/SHMStats.h"
 
 #include "LinuxSyscalls/Types.h"
 #include "LinuxSyscalls/Seccomp/SeccompEmulator.h"
@@ -123,7 +123,7 @@ public:
 
   ~ThreadManager();
 
-  class StatAlloc final : public FEX::Profiler::StatAllocBase {
+  class StatAlloc final : public FEX::SHMStats::StatAllocBase {
   public:
     StatAlloc();
 
@@ -132,8 +132,8 @@ public:
 
     void CleanupForExit();
 
-    FEXCore::Profiler::ThreadStats* AllocateSlot(uint32_t TID);
-    void DeallocateSlot(FEXCore::Profiler::ThreadStats* AllocatedSlot);
+    FEXCore::SHMStats::ThreadStats* AllocateSlot(uint32_t TID);
+    void DeallocateSlot(FEXCore::SHMStats::ThreadStats* AllocatedSlot);
 
   private:
     void Initialize();
