+++ Source/Windows/Common/SHMStats.cpp
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-#include "Windows/Common/Profiler.h"
+#include "Windows/Common/SHMStats.h"
 
 #include <FEXCore/fextl/fmt.h>
 #include <FEXCore/Utils/LogManager.h>
@@ -40,7 +40,7 @@ uint32_t StatAlloc::FrontendAllocateSlots(uint32_t NewSize) {
   return CurrentSize;
 }
 
-StatAlloc::StatAlloc(FEXCore::Profiler::AppType AppType) {
+StatAlloc::StatAlloc(FEXCore::SHMStats::AppType AppType) {
   // Try wine+fex magic path.
 
   {
